+++ CHANGES.rst
@@ -66,6 +66,11 @@ astropy.units
 
 astropy.utils
 ^^^^^^^^^^^^^
+astropy.units
+^^^^^^^^^^^^^
+
+- Fixed a bug that prevented ``EarthLocation`` from being initialized with
+  numpy >=1.17. [#8849]
 
 astropy.visualization
 ^^^^^^^^^^^^^^^^^^^^^